refactor(Answer:5): app component with todo-list-item component
