+++ apps/angular/crud/src/app/app.component.ts
@@ -1,16 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit, WritableSignal } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatDialog } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { LoadingDialog } from './dialogs';
 import { Todo } from './models';
 import { TodoService } from './services';
 
 @Component({
   standalone: true,
-  imports: [CommonModule, MatCardModule, MatCheckboxModule],
+  imports: [CommonModule, MatCardModule],
   selector: 'app-root',
   template: `
     <div class="container">
@@ -19,16 +15,7 @@ import { TodoService } from './services';
           <ul>
             @for (todo of todos(); track todo.id) {
               <li>
-                <div class="todo-list-item">
-                  <div>
-                    <mat-checkbox></mat-checkbox>
-                    <span>{{ todo.title }}</span>
-                  </div>
-                  <div>
-                    <button (click)="updateTodo(todo)">Update</button>
-                    <button (click)="deleteTodo(todo)">Delete</button>
-                  </div>
-                </div>
+                <app-todo-list-item></app-todo-list-item>
               </li>
             }
           </ul>
@@ -46,66 +33,14 @@ import { TodoService } from './services';
       list-style-type: none;
       padding-inline-start: 0px;
     }
-
-    .todo-list-item {
-      display: flex;
-      justify-content: space-between;
-      gap: 10px;
-
-      & > div {
-        display: flex;
-        align-items: center;
-        height: fit-content;
-        gap: 10px;
-      }
-    }
   `,
 })
 export class AppComponent implements OnInit {
   todos: WritableSignal<Todo[]> = this.todoService.todos;
 
-  constructor(
-    private readonly todoService: TodoService,
-    private readonly snackbar: MatSnackBar,
-    public dialog: MatDialog,
-  ) {}
+  constructor(private readonly todoService: TodoService) {}
 
   ngOnInit(): void {
     this.todoService.getAll();
-  }
-
-  updateTodo(todo: Todo): void {
-    this.dialog.open(LoadingDialog);
-    this.todoService.update(todo).subscribe(
-      (updatedTodo: Todo) => this.updateArrayItem(this.todos(), updatedTodo),
-      (error) => this.showErrorMessage(),
-    );
-  }
-
-  deleteTodo(todo: Todo): void {
-    this.dialog.open(LoadingDialog);
-    this.todoService.delete(todo).subscribe(
-      () => this.removeArrayItem(this.todos(), todo),
-      (error) => this.showErrorMessage(),
-    );
-  }
-
-  showErrorMessage(): void {
-    this.dialog.closeAll();
-    this.snackbar.open('Error occured', undefined, { duration: 3000 });
-  }
-
-  updateArrayItem(array: Todo[], updatedTodo: Todo): void {
-    if (this.dialog.openDialogs) {
-      this.dialog.closeAll();
-    }
-    this.todos.set(Todo.updateItemInArray(array, updatedTodo));
-  }
-
-  removeArrayItem(array: Todo[], deleteTodo: Todo): void {
-    if (this.dialog.openDialogs) {
-      this.dialog.closeAll();
-    }
-    this.todos.set(Todo.removeItemFromArray(array, deleteTodo));
   }
 }
